worker/requirements.txt
google-cloud-pubsub==0.30.1
google-cloud-firestore==0.28.0
google-cloud-trace==0.17.0
opencensus==0.1.3
wrapt==1.10.11